requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.108.0
h11==0.14.0
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.4
pandas==2.0.3
pydantic==2.5.3
pydantic-core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
typing-extensions==4.9.0
tzdata==2023.4
uvicorn==0.25.0
